Project structure fix, routes fix, styling bux fix
There are plenty things to improve but at least, it is working.
