+++ frontend/package.json
@@ -53,6 +53,5 @@
     "jest-expo": "^36.0.1",
     "react-native-scripts": "^2.0.1",
     "react-test-renderer": "^16.2.0"
-  },
-  "proxy": "http://localhost:8080/"
+  }
 }
