+++ frontend/src/components/Guestpage/GuestpageContent.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Body from './body';
+
+//+ plus appbar
+class GuestpageContent extends React.Component {
+  render() {
+    return (
+      <div className="guestpanel">
+        <Body />
+      </div>
+    );
+  }
+}
+
+export default GuestpageContent;
