+++ frontend/src/components/Guestpage/body.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import './body.css';
-import check from '../../pictures/check.png';
-import flower from '../../pictures/flower.png';
-import choiceForm from '../../pictures/choiceForm.png';
-import statistics from '../../pictures/statistics.jpg';
-import fiszka from '../../pictures/fiszka.png';
+import './startBody.css';
+import check from '../../images/check.png';
+import flower from '../../images/flower.png';
+import choiceForm from '../../images/choiceForm.png';
+import statistics from '../../images/statistics.jpg';
+import fiszka from '../../images/fiszka.png';
 import { Redirect } from 'react-router-dom';
 
 class Body extends React.Component {
@@ -48,7 +48,7 @@ class Body extends React.Component {
           </div>
 
           <div className="pic">
-            <img src={fiszka}></img>
+            <img src={fiszka} alt={'obrazek'}></img>
           </div>
         </div>
 
@@ -59,7 +59,7 @@ class Body extends React.Component {
           </div>
 
           <div className="pic">
-            <img src={statistics}></img>
+            <img src={statistics} alt={'obrazek'}></img>
           </div>
         </div>
