+++ frontend/src/components/Guestpage/index.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import GuestpageContent from './GuestpageContent';
+
+const Guestpage = () => {
+  return (
+    <div>
+      <GuestpageContent />;
+    </div>
+  );
+};
+
+export default Guestpage;
