+++ frontend/src/components/Homepage/HomepageContent.jsx
@@ -1,16 +1,14 @@
 import React, { Component } from 'react';
 import Body from './body';
-import NavBar from './NavBar';
 
-class Homepage extends Component {
+class HomepageContent extends Component {
   render() {
     return (
       <div className="homepanel">
-        <NavBar />
         <Body />
       </div>
     );
   }
 }
 
-export default Homepage;
+export default HomepageContent;
