+++ frontend/src/components/Login/PasswordReset.jsx
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+
+class PasswordReset extends Component {
+  render() {
+    return (
+      <div>
+        PasswordReset<br></br>
+        PasswordReset<br></br>
+        PasswordReset<br></br>
+        PasswordReset<br></br>
+      </div>
+    );
+  }
+}
+
+export default PasswordReset;
