--- frontend/src/components/Login/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import LoginPanel from './loginPanel';
-
-const LoginContent = () => {
-  return (
-    <div>
-      <LoginPanel />
-    </div>
-  );
-};
-
-export default LoginContent;
