--- frontend/src/components/Login/passwordReset.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-class PasswordReset extends Component {
-  render() {
-    return <div>PasswordReset</div>;
-  }
-}
-
-export default PasswordReset;
