+++ frontend/src/components/Register/RegisterPanel.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import './register.css';
-import logo from './logo.png';
+import logo from '../../images/logo.png';
 import { Link } from 'react-router-dom';
 
 class RegisterPanel extends React.Component {
