--- frontend/src/components/Register/index.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import RegisterPanel from './registerPanel';
-
-const RegisterContent = () => {
-  return (
-    <div>
-      <RegisterPanel />
-    </div>
-  );
-};
-
-export default RegisterContent;
