+++ frontend/src/components/Register/index.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import RegisterPanel from './RegisterPanel';
+
+const Register = () => {
+  return (
+    <div>
+      <RegisterPanel />
+    </div>
+  );
+};
+
+export default Register;
