--- frontend/src/components/Startpage/startpage.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import NavBar from './NavBar';
-import Body from './body';
-
-//+ plus appbar
-class startpage extends React.Component {
-  render() {
-    return (
-      <div className="homepanel">
-        <NavBar />
-
-        <Body />
-      </div>
-    );
-  }
-}
-
-export default startpage;
