+++ frontend/src/images/bugIcon.png
