--- frontend/src/views/Login/index.js
@@ -1,8 +1,0 @@
-import LoginContent from '../../components/Login';
-import React from 'react';
-
-const Login = () => {
-  return <LoginContent />;
-};
-
-export default Login;
