--- frontend/src/views/Register/index.js
@@ -1,8 +1,0 @@
-import RegisterContent from '../../components/Register';
-import React from 'react';
-
-const Register = () => {
-  return <RegisterContent />;
-};
-
-export default Register;
