--- frontend/src/views/Startpage/index.js
@@ -1,8 +1,0 @@
-import StartContent from '../../components/Register';
-import React from 'react';
-
-const Start = () => {
-  return <StartContent />;
-};
-
-export default Start;
